refactor/smallish-refactoring-and-documentation-improvements
