+++ CHANGELOG.md
@@ -1,0 +1,19 @@
+Version 1.0.2 (2024-04-03)
+==========================
+
+- Cleaned some things up mostly and changed documentation.
+- For the next version I'm planning to add more tests and making the github prettier. I will also make what I got even neater and then the version after that I aim to add more graph algorithms. After that I can start creating some benchmarks to see how much/if there is a significant speedup in some situations compared to alternatives and when that happen.
+
+Version 1.0.1 (2024-04-03)
+==========================
+
+- Some very minor fixes
+
+Version 1.0.0 (2024-04-03)
+==========================
+
+Breaking changes
+----------------
+
+- Everything is new!
+
