+++ Cargo.toml
@@ -5,7 +5,7 @@ documentation = "https://docs.rs/fast-graph"
 readme = "README.md"
 homepage = "https://github.com/henke443/fast-graph"
 repository = "https://github.com/henke443/fast-graph"
-version = "1.0.1"
+version = "1.0.2"
 edition = "2021"
 license = "Apache-2.0"
 categories = ["data-structures", "database", "mathematics", "science::geo", "visualization"]
