+++ src/categories.rs
@@ -33,6 +33,7 @@
 use crate::*;
 #[cfg(feature = "hashbrown")]
 use hashbrown::HashMap;
+
 #[cfg(not(feature = "hashbrown"))]
 use std::collections::HashMap;
 use thiserror::Error;
